+++ apps/deployex_web/lib/deployex_web/live/components/application_card.ex
@@ -135,6 +135,7 @@ defmodule DeployexWeb.Components.ApplicationCard do
                   <% end %>
                 </div>
               </div>
+              <.served_urls id={@application.sname} urls={@application.urls} />
             </div>
           </div>
           <div>